+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [x.y.z](https://github.com/sushichop/Puppy/releases/tag/x.y.z) (yyyy-mm-dd)
+
+- Instantiate `DispatchQueue` and add an argument named of `asynchronous`. [#9](https://github.com/sushichop/Puppy/pull/9)
+
 ## [0.1.2](https://github.com/sushichop/Puppy/releases/tag/0.1.2) (2020-12-05)
 
 - Support new API about FileHandle. [#7](https://github.com/sushichop/Puppy/pull/7)